# openapi.yaml
...
openapi: 3.1.0
info:
  title: VintaSend Managed Templates API
  version: 1.0.0
  description: 'HTTP contract for managing VintaSend notification templates: their
    versions, their status lifecycle, and previewing a version before it is published.'
paths:
...
  /api/v1/tags:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_list_tags
      summary: List Tags
      parameters:
      - in: query
        name: page
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
        required: false
      - in: query
        name: pageSize
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
        required: false
      - in: query
        name: status
        schema:
          anyOf:
          - items:
              enum:
              - active
              - archived
              type: string
            type: array
          - type: 'null'
          title: Status
        required: false
      - in: query
        name: search
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
        required: false
      - in: query
        name: tenant
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant
        required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ManagedTemplateTagOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'List tags, newest filter first: by status, by a text search, or
        by tenant.


        A tag picker wants ``?status=active``: archived tags are still attached to
        the templates

        carrying them and can still be filtered on, they are simply no longer offered.'
      tags:
      - tags
      security:
      - ApiKeyAuth: []
    post:
      operationId: vintasend_templates_management_api_templates_manager_api_create_tag
      summary: Create Tag
      parameters: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateTagOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Define a tag ahead of any template using it.


        Tagging a template creates missing tags on its own, so this is for the case
        where a

        collision with an existing tag is worth hearing about -- it is a 409 here,
        where tagging

        a template would silently resolve to the tag already there.'
      tags:
      - tags
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagBody'
        required: true
      security:
      - ApiKeyAuth: []
  /api/v1/tags/{slug}:
    get:
      operationId: vintasend_templates_management_api_templates_manager_api_get_tag
      summary: Get Tag
      parameters:
      - in: path
        name: slug
        schema:
          title: Slug
          type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateTagOut_'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: One tag. The path accepts the tag's slug or the text it was created
        from.
      tags:
      - tags
      security:
      - ApiKeyAuth: []
    patch:
      operationId: vintasend_templates_management_api_templates_manager_api_update_tag
      summary: Update Tag
      parameters:
      - in: path
        name: slug
        schema:
          title: Slug
          type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateTagOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Rename a tag. Its slug is regenerated, so its URL changes.


        The templates carrying the tag keep it, but a stored filter naming the old
        slug stops

        matching -- read ``slug`` off the response and store that. Renaming a tag
        onto another

        tag''s text is allowed and yields a ``-2`` suffix: two tags may legitimately
        read the

        same, and the slug is what tells them apart.'
      tags:
      - tags
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagBody'
        required: true
      security:
      - ApiKeyAuth: []
    delete:
      operationId: vintasend_templates_management_api_templates_manager_api_delete_tag
      summary: Delete Tag
      parameters:
      - in: path
        name: slug
        schema:
          title: Slug
          type: string
        required: true
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Delete a tag, removing the label from every template carrying
        it.


        Not reversible, and the templates lose the label. Archive instead when the
        tag should

        stop being offered but the templates should keep it.'
      tags:
      - tags
      security:
      - ApiKeyAuth: []
  /api/v1/tags/{slug}/archive:
    post:
      operationId: vintasend_templates_management_api_templates_manager_api_archive_tag
      summary: Archive Tag
      parameters:
      - in: path
        name: slug
        schema:
          title: Slug
          type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateTagOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Retire a tag from the pickers without touching the templates carrying
        it.


        Filtering by an archived tag still returns those templates -- archiving hides
        the tag

        from ``?status=active``, it does not hide the templates.'
      tags:
      - tags
      security:
      - ApiKeyAuth: []
  /api/v1/tags/{slug}/restore:
    post:
      operationId: vintasend_templates_management_api_templates_manager_api_restore_tag
      summary: Restore Tag
      parameters:
      - in: path
        name: slug
        schema:
          title: Slug
          type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse_ManagedTemplateTagOut_'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      description: 'Put an archived tag back on offer.


        Unlike an archived template version -- terminal, because reviving one would
        rewrite what

        its audit trail says happened -- a tag carries no history to contradict, so
        archiving one

        is reversible.'
      tags:
      - tags
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
...
    ApiErrorBody:
      properties:
        code:
          enum:
          - BAD_REQUEST
          - UNAUTHORIZED
          - NOT_FOUND
          - CONFLICT
          - INVALID_STATUS_TRANSITION
          - PREVIEW_UNAVAILABLE
          - TEMPLATE_COMPOSITION_ERROR
          - INTERNAL_ERROR
          title: Code
          type: string
        message:
          title: Message
          type: string
        details:
          anyOf:
          - $ref: '#/components/schemas/JsonValue'
          - type: 'null'
      required:
      - code
      - message
      title: ApiErrorBody
      type: object
    ApiErrorResponse:
      description: Error envelope returned with every non-2xx response.
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorBody'
      required:
      - error
      title: ApiErrorResponse
      type: object
    JsonValue: {}
...
    ManagedTemplateTagOut:
      description: 'A label attached to any number of template versions.


        ``slug`` is the tag''s identity: it is normalized from ``text``, unique store-wide,
        and

        what the ``includesAllTags`` / ``includesAnyOfTags`` filters match on. A client
        that

        stores a tag reference should store the slug, and re-read it after a rename
        -- editing

        ``text`` regenerates the slug, so the old one stops matching.'
      properties:
        id:
          title: Id
          type: string
        text:
          title: Text
          type: string
        slug:
          title: Slug
          type: string
        status:
          enum:
          - active
          - archived
          title: Status
          type: string
        tenant:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
        updatedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedat
      required:
      - id
      - text
      - slug
      - status
      - tenant
      - createdAt
      - updatedAt
      title: ManagedTemplateTagOut
      type: object
...
    PaginatedResponse_ManagedTemplateTagOut_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ManagedTemplateTagOut'
          title: Data
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        hasMore:
          title: Hasmore
          type: boolean
      required:
      - data
      - page
      - pageSize
      - hasMore
      title: PaginatedResponse[ManagedTemplateTagOut]
      type: object
    DataResponse_ManagedTemplateTagOut_:
      properties:
        data:
          $ref: '#/components/schemas/ManagedTemplateTagOut'
      required:
      - data
      title: DataResponse[ManagedTemplateTagOut]
      type: object
    CreateTagBody:
      description: 'Body of ``POST /api/v1/tags``.


        The slug is derived from ``text`` by the library and is not a client''s to
        set: it is the

        tag''s identity, and a client-supplied one could name a tag no other caller
        would produce.'
      properties:
        text:
          maxLength: 255
          minLength: 1
          title: Text
          type: string
        tenant:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Tenant
      required:
      - text
      title: CreateTagBody
      type: object
    UpdateTagBody:
      description: 'Body of ``PATCH /api/v1/tags/{slug}``, which renames a tag.


        The slug is regenerated from the new text, so the tag''s URL changes and a
        stored filter

        naming the old slug stops matching. The templates carrying the tag keep it.'
      properties:
        text:
          maxLength: 255
          minLength: 1
          title: Text
          type: string
      required:
      - text
      title: UpdateTagBody
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
servers: []